# Rework checkpoint reading from database to avoid quadratic queries with number of assets